TDS-396: refactoring based on feedback
