+++ service/src/main/java/tds/exam/configuration/data/DataSourceConfiguration.java
@@ -1,13 +1,15 @@
 package tds.exam.configuration.data;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import javax.sql.DataSource;
+
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceBuilder;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
-
-import javax.sql.DataSource;
+import org.springframework.jdbc.datasource.DataSourceTransactionManager;
+import org.springframework.transaction.PlatformTransactionManager;
 
 /**
  * Configuration for DataSources.  Queries
@@ -38,5 +40,12 @@ public class DataSourceConfiguration {
     @Bean(name = "queryJdbcTemplate")
     public NamedParameterJdbcTemplate queryJdbcTemplate() {
         return new NamedParameterJdbcTemplate(queriesDataSource());
+    }
+
+    @Bean
+    public PlatformTransactionManager transactionManager(@Qualifier("commandDataSource") DataSource dataSource) {
+        final DataSourceTransactionManager dataSourceTransactionManager = new DataSourceTransactionManager();
+        dataSourceTransactionManager.setDataSource(dataSource);
+        return dataSourceTransactionManager;
     }
 }
