+++ service/src/main/java/tds/exam/configuration/web/WebConfiguration.java
@@ -1,0 +1,18 @@
+package tds.exam.configuration.web;
+
+
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+/**
+ * Configuration for REST templates.
+ */
+@Configuration
+public class WebConfiguration {
+    @Bean
+    public RestTemplate restTemplate(RestTemplateBuilder builder) {
+        return builder.build();
+    }
+}
