--- service/src/test/java/tds/exam/config/WebConfiguration.java
@@ -1,15 +1,0 @@
-package tds.exam.config;
-
-
-    import org.springframework.boot.web.client.RestTemplateBuilder;
-    import org.springframework.context.annotation.Bean;
-    import org.springframework.context.annotation.Configuration;
-    import org.springframework.web.client.RestTemplate;
-
-@Configuration
-public class WebConfiguration {
-    @Bean
-    public RestTemplate restTemplate(RestTemplateBuilder builder) {
-        return builder.build();
-    }
-}
